Fill in more of the Camera Class code from Nicla plus
There were several of the methods defined in the main camera
class that were not implemented.  That is that the compiler
would complete, but the linker would fail.

It was a simple case of looking at the Camera code in the
Arduino Nicla Vision, to find the implementations.

I extended the class slightly, as most if not all of the
cameras, already had a debug method, so I added it
to the Camera class.  I also added extended support
for allowing you to print out the current state
of the camera registers.

Some of the classes in particular HM01B0 and HM0360 had
not implemented the camera vertical flip and Horizontal
mirror, so put that in.

Also put in the change to add 32 bytes to the malloc
for the frame buffer as the code to make sure it was
on a 32 byte boundary could cause memory corruption